Updated WindowsManager#updateWindowBounds
Previously, when setting the bounds of a window, the inner Window component received the new dimensions, causing the enclosing NativeWindow to resize as a result. This yielded native windows slightly larger than the requested size, in order to accommodate the native chrome.

Now I changed that, the way that the new size is received by the NativeWindow directly, and (only after) the inner Window component receives the same size minus the native chrome, if any. This results in slightly smaller real estate available to the window content, but I believe it is more acceptable this way.